Scripts update : email + alerts
Script can now send mail to one or more email adress if user exceeds 1 Gigabyte of disk usage in 'directorytosort' and sub-directories.

A Windows Alert box will spawn too, that indicates which user it is.

(had to removed 1GB files from commit, just put some in directorytosort to test it)